requirements.txt
absl-py==1.3.0
aiofiles==22.1.0
anyio==3.6.2
astunparse==1.6.3
autopep8==2.0.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dill==0.3.6
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.95.1
flatbuffers==22.11.23
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h11==0.14.0
h5py==3.7.0
httpcore==0.16.1
httptools==0.5.0
httpx==0.23.1
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
motor==3.1.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
orjson==3.8.2
packaging==21.3
Pillow==9.3.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pydantic==1.10.2
pymongo==4.3.3
pyparsing==3.0.9
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
termcolor==2.1.1
threadpoolctl==3.1.0
tomli==2.0.1
typing_extensions==4.4.0
ujson==5.5.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
